+++ src/routes/userRoutes.ts
@@ -1,10 +1,10 @@
 import { Router } from 'express';
-import { getUsers, createUser } from '../controllers/userController';
+import { getUsers, createUser, updateUser } from '../controllers/userController';
 
 const router = Router();
 
 router.get('/', getUsers);
 router.post('/', createUser);
+router.put('/:id', updateUser);
 
 export default router;
-
